Put query points in own file, remove smoothing an use fixed query points instead.
