+++ src/sample_expression.rs
@@ -24,6 +24,7 @@ use crate::common::Outdir;
 use crate::errors::Error;
 use crate::preprocess::Preprocessing;
 use crate::prob_distribution_2d::ProbDistribution2d;
+use crate::query_points;
 
 pub(crate) fn sample_expression(
     preprocessing: &Path,
@@ -34,6 +35,7 @@ pub(crate) fn sample_expression(
 ) -> Result<()> {
     
     let preprocessing = Preprocessing::from_path(preprocessing)?;
+    let sample_ids = preprocessing.scale_factors().keys().cloned().collect::<Vec<_>>();
     let prior = preprocessing.prior()?;
     let mean_disp_estimates =
         preprocessing
@@ -42,10 +44,6 @@ pub(crate) fn sample_expression(
             .ok_or(Error::UnknownSampleId {
                 sample_id: sample_id.to_owned(),
             })?;
-
-    let query_points = preprocessing.query_points();
-    // let mu_ik_points = Vec::<f64>::new();//query_points.all_mu_ik();
-    // let start_points_theta_i = Vec::<f64>::new(); //query_points.thetas();
 
     let mut s_j = preprocessing
         .scale_factors()
@@ -56,22 +54,28 @@ pub(crate) fn sample_expression(
     // let temp: Vec<_>= mean_disp_estimates.means().iter().enumerate().skip(190432).collect();
     // println!("mean_disp_estimates {:?}", temp);
 
-    let file = File::open("/vol/nano/bayesian-diff-exp-analysis/texp-evaluation/estimated_dispersion.csv")?;
-    let mut rdr = csv::Reader::from_reader(file);
-    let mut thetas = Vec::<f64>::new();
-    for result in rdr.records() {
-        let record = result?;
-        let dispersion: f64 = record[1].parse().unwrap();
-        thetas.push(dispersion);
-    }
+    let query_points = query_points::calc_query_points(c, preprocessing.mean_disp_estimates().clone(), sample_ids, preprocessing.feature_ids().clone());
+    // let query_points = preprocessing.query_points();
+    // let mu_ik_points = Vec::<f64>::new();//query_points.all_mu_ik();
+    // let start_points_theta_i = Vec::<f64>::new(); //query_points.thetas();
+
+    // let file = File::open("/vol/nano/bayesian-diff-exp-analysis/texp-evaluation/estimated_dispersion.csv")?;
+    // let mut rdr = csv::Reader::from_reader(file);
+    // let mut thetas = Vec::<f64>::new();
+    // for result in rdr.records() {
+    //     let record = result?;
+    //     let dispersion: f64 = record[1].parse().unwrap();
+    //     thetas.push(dispersion);
+    // }
 
 
 
     let mut feature_ids: Vec<_> = preprocessing.feature_ids().iter().enumerate().skip(190432).collect();
     // println!("{:?} features", feature_ids.len());
     // println!("{:?}", feature_ids);
-    let subsampled_ids = vec!["ERCC-00130","ERCC-00004", "ERCC-00136", "ERCC-00096", "ERCC-00171", "ERCC-00009",
-    "ERCC-00074", "ERCC-00113", "ERCC-00145", "ERCC-00002", "ERCC-00046", "ERCC-00003"];
+    // let subsampled_ids = vec!["ERCC-00147","ERCC-00117", "ERCC-00016", "ERCC-00086", "ERCC-0061", "ERCC-00048"];
+    let subsampled_ids = vec!["ERCC-00130","ERCC-00004", "ERCC-00136", "ERCC-00096", "ERCC-00074", "ERCC-00002"];
+    // ,    "ERCC-00171", "ERCC-00113", "ERCC-00145", "ERCC-00002", "ERCC-00046", "ERCC-00003", "ERCC-00113"];
 
     let out_dir = Outdir::create(out_dir_path)?;
     // feature_ids.truncate(10000);
@@ -81,16 +85,19 @@ pub(crate) fn sample_expression(
             
             // if subsampled_ids.contains(&feature_id.as_str()) {
             // if feature_id.as_str() == "ERCC-00108" {   
+            // prit start time of feature
+            let time1= std::time::SystemTime::now();
+            // println!("feature {:?} {:?} started at {:?}", i, feature_id, time1 );
 
             // println!("\n--------------feature {:?} {:?}", i, feature_id);
-            let mut output = out_dir_path.to_path_buf();
-            output.push(feature_id);
-            output.set_extension("csv");
-            let mut wtr = csv::Writer::from_path(output)?;
-            wtr.serialize(("mu_ik", "probability")).unwrap();
+            // let mut output = out_dir_path.to_path_buf();
+            // output.push(feature_id);
+            // output.set_extension("csv");
+            // let mut wtr = csv::Writer::from_path(output)?;
+            // wtr.serialize(("mu_ik", "probability")).unwrap();
 
             let d_ij = mean_disp_estimates.means()[*i]; //TODO Do we need group mean mu_ik instead of sample mean
-            // println!("d_ij {:?}", d_ij);
+            // println!("{:?} d_ij {:?}", feature_id, d_ij / s_j);
             // METHOD: If the per-sample dispersion is unknown, fall back to a mean interpolated from the other samples.
             let mut t_ij = if let Some(t_ij) = mean_disp_estimates.dispersions()[*i] {
                 t_ij
@@ -142,21 +149,24 @@ pub(crate) fn sample_expression(
                     s_j,
                     epsilon,
                 );
-                if t == 0.01 {
-                    wtr.serialize((m, prob.exp())).unwrap();
-                    // if m < 5. {
-                    //     println!("mu {:?}, prob {:?}", m, prob.exp());
-                    // }
-                }
+                // if t == 0.01 {
+                //     wtr.serialize((m, prob.exp())).unwrap();
+                //     // if m < 5. {
+                //     //     println!("mu {:?}, prob {:?}", m, prob.exp());
+                //     // }
+                // }
                 prob
             };
             let mu_ik_points = query_points.get(&feature_id.to_string()).unwrap().all_mu_ik();
             let start_points_theta_i = query_points.get(&feature_id.to_string()).unwrap().thetas();
-            // println!("mu_ik_points {:?} {:?}", mu_ik_points[0], mu_ik_points[mu_ik_points.len()-1]);
+            // println!("{:?} #mu_ik {:?}, #theta_i {:?}",feature_id,  mu_ik_points.len(), start_points_theta_i.len());
+            // println!("{:?} mu_ik_points {:?} {:?}",feature_id, mu_ik_points[0], mu_ik_points[mu_ik_points.len()-1]);
             // println!("start_points_theta_i {:?}", start_points_theta_i);
-            likelihoods.insert_grid(mu_ik_points.clone(), start_points_theta_i.clone(), calc_prob);
+            likelihoods.insert_grid(d_ij/s_j, mu_ik_points.clone(), start_points_theta_i.clone(), calc_prob);
 
             out_dir.serialize_value(feature_id, likelihoods)?;
+            let time2= std::time::SystemTime::now();
+            println!("sample {} feature {:?} {:?} finished in duration {:?}",sample_id, i, feature_id, time2.duration_since(time1).unwrap());
             // }
             Ok(())
             
@@ -198,8 +208,8 @@ fn prob_mu_ik_theta_i_x(
     // println!("x {:?}, d_ij {:?}, mu_ik {:?}, t_ij {:?}, theta_i {:?}, s_j {:?}", x, d_ij, mu_ik, t_ij, theta_i, s_j);
     //METHOD /s_j?? mu_ik*s_j würde einluss von t_ij ändern, weil mu geändert wird.
     let left = neg_binom(d_ij/s_j, x, t_ij);
-    // let right = neg_binom(x, mu_ik, theta_i);
-    let right =  LogProb::from(Poisson::new(mu_ik).unwrap().ln_pmf(x as u64));
+    let right = neg_binom(x, mu_ik, theta_i);
+    // let right =  LogProb::from(Poisson::new(mu_ik).unwrap().ln_pmf(x as u64));
     let result = left + right; 
     // if theta_i == 0.01 && mu_ik < 5.{
     //     println!("x {:?}, mu_ik {:?}, left {:?}, right {:?}, result {:?}", x, mu_ik, left.exp(), right.exp(), result.exp());
@@ -225,37 +235,25 @@ fn likelihood_mu_ik_theta_i(
     }
     let mut max_prob = LogProb::ln_zero();
     let mut result = LogProb::ln_zero();
-
-    // let density = |_, x: f64| prob_mu_ik_theta_i_x(x, d_ij, mu_ik, t_ij, theta_i, s_j);
-
-    // result= LogProb::ln_simpsons_integrate_exp(
-    //         density,
-    //         0.,
-    //         300.,
-    //         1001,
-    //     );
-
-    // result.ln_add_exp(LogProb::ln_simpsons_integrate_exp(
-    //         density,
-    //         300.,
-    //         10000.,
-    //         3001,
-    //     ));
+    // println!("mu_ik {:?} d_ij {:?}", mu_ik, d_ij/s_j);
     let mut x : f64 = 0.;
+    
     loop {
         let calced_prob = prob_mu_ik_theta_i_x(x as f64, d_ij, mu_ik, t_ij, theta_i, s_j);
         if calced_prob > max_prob{
             max_prob = calced_prob;
         }       
         result = result.ln_add_exp(calced_prob);
-        // if x > 0. && calced_prob.exp() == 0.{
-        //     break;
-        // }
-        if x > 250. && calced_prob - max_prob < LogProb(0.000001_f64.ln()) {
+
+        if (x > 200. && calced_prob - max_prob < LogProb(0.001_f64.ln())) {
             break;
         }
+        // if x % 1000. == 0. {
+        //     println!("x {:?}, calced_prob {:?}, max_prob {:?} diff {:?}", x, calced_prob.exp(), max_prob.exp(), (calced_prob - max_prob).exp());
+        // }
         x = x + 1.;
     }
+    // println!("mu_ik {:?} x {:?}", mu_ik, x); 
     // if theta_i < 0.3 {
     //     println!("mu_ik {:?}, theta_i {:?}, x {:?}, result {:?}, max_prob {:?}", mu_ik, theta_i, x, result, max_prob);
     // }
